cli.py: Suppress a warning message from fedmsg
fedmsg's routing policy mechanism is meant to allow binding particular
messages to the crypto certificates for particular servers in the
infrastructure. Setting this up outside of Fedora infrastructure would
be difficult, so we just filter out the nag messages that occur when
no routing policy is set, but messages aren't actually blocked.