Set up alembic for database migrations
Add an alembic configuration to allow migrations as the database
changes. Migrations are applied automatically when flatpak-status
is run - since everything in the database is a cache, the worst that
can happen is that it's necessary to remove the database and let
the database be refilled from Bodhi and Koji.